web/requirements.txt
django>=3.2,<4.2
gunicorn
requests
django-dotenv
psycopg2-binary
django-storages
boto3
django-rest-framework
django-cors-headers
pandas